+++ examples/demo20/src/scatter/App.vue
@@ -28,6 +28,7 @@ export default {
     getChart(this.$refs.canvas, echarts, {
       width: this.width,
       height: this.height,
+      devicePixelRatio: systemInfo.devicePixelRatio,
     }).then(this.initChart)
   },
   methods: {